requirements.txt
# Web Framework
Flask

# Excel Processing
pandas
openpyxl

# Git Operations
GitPython

# Task Scheduling
schedule
